Moves most (but not yet all) hydration code paths from deploy
