+++ src/deploy/lambda-one/deploy.js
@@ -4,9 +4,9 @@ var waterfall = require('run-waterfall')
 
 // local deps
 var _getName = require('../helpers/get-function-name')
-var uploadZip = require('./06-upload-zip')
-var afterDeploy = require('./07-after-deploy')
-var done = require('./08-done')
+var uploadZip = require('./03-upload-zip')
+var afterDeploy = require('./04-after-deploy')
+var done = require('./05-done')
 
 module.exports = function deploy(params, callback) {
