+++ src/deploy/lambda-one/index.js
@@ -20,6 +20,7 @@ module.exports = function deployOne(params, callback) {
   })
 
   let arc = params.arc
+  params.hydrateDeps = true
 
   const _prep = prep.bind({}, params)
   const _deploy = deploy.bind({}, params)
